ClickPattern - Add new check [Not Done/Beta]
This check will attempt to find patterns in a players click. It calculates the time between each click, and checks if the difference is the same or near the last click difference. This can be very effective against combat cheats that do not have any cps deviation. 

ToDo: Average difference: Another clicking pattern check by comparing the average click difference and seeing if they are all relatively the same.