Rework Criticals check
Attempt to look for silent jumps by comparing the players fallDistance and the fallDistance calculated by NoFall. Only check if that fall distance is below/equal to the fall disatnce set in the config. Use the jumpPhase to help determine if it was a silent jump as well

This method should reduce the false positives from the old check, by now focusing on silent jumps rather than just crit hits within a fall distance.