[Bleeding] Changes (read on)
Sf:
* Take care of the last left-over from asofold -> Xaw3ep/CaptainObvious transition: add limits for moving in bubble columns, distinguish drag/push, deny ascension if getting dragged or descend if getting pushed.
* Bunny + headbang: add velocity on certain transitions. Modelling is too much of a hassle here.
* Alter hAcc limit slightly;
* Confine liquid workarounds a bit more;
* Modify speed limits when moving on surface while bouncing on bubble column;
* Deny ascension in liquid if sneaking;
* Waterwalk: fix potential "bypass" by checking with lower hDistance (using 2 moves to prevent false positives).

[Experimental] Sf velocity:
If strict invalidation is set to true, *ALL* horizontal velocity is cleared (queued as well, not only active) with no hDistance. Effects on actual cheats remain to be seen, but this should prevent cheaters from turning on anti-knockback hacks to conveniently collect velocity entries for a one time cheat attempt (longjump). Coarse, would be better to get the location at which a velocity entry was collected/queued and compare it to the location at which the entry was used instead.

Sf, Lava:
* Lava, horizontal move: bunnyhopping on//near lava's 6th level doesn’t slow the player down enough, it also won’t break the player’s sprint like water does. This will cause false positive as NCP will apply the liquid speed limit regardless. Also the bunnyhop mechanic won’t apply due to the change of envelope. Occasional false positives might still appear but it should be *a lot* better now.
* Lava, vertical move:
Add a vDistRel case, rather rare, was able to observe it just one time (not sure if it happens with legacy clients);
Add LiquidWorkaround case: jumping with normal gravity on the lowest lava level.

Block flags:
* Add bouncing flag to beds only from MC 1.12 and on;
* Remove the F_HEIGHT_8SIM_DEC flag from lava only from MC 1.13 and on, re-add it back from 1.16 because Mojang reverted lava movement back to pre 1.13 behaviour; however at the lowest level one can now jump up and down normally: resolved with adding the flag back up.

Blockplace:
* Re-balance Improbable feeding.

Missing:
* Launch effect for bubble streams;
* Sneaking modifier: Odd false positives, not sure what's going on here.
* Bubble columns are still horrendously buggy.