Attempt to get the client's protocol ID thorugh CompatNoCheatPlus
This should be done within NoCheatPlus rather.
It is also a bare-bone attempt: you'll need to look up the protocol ID table to know which version the player is on.
Later, some kind of translation method (from protocol to launcher version) might be done.

(PRs to improve this are most welcomed :p)